Create home page and add first recipe
The home page will be altered as new recipes are added. These recipes are organized in an unsorted list.
The recipe page consists of a picture followed by a description, ingredients and steps. The original recipe is linked at the bottom.
Images and recipes are stored in separate directories. All relative.
Made using only html.